requirements.txt
discord.py==2.1.*
python-dotenv==0.21.*
openai==1.55.*
PyYAML==6.0
dacite==1.6.*
numpy
